refactor(dex): break the measured-execution module cycle via a leaf type
check:shared-cycles flagged dex-liquidity/types -> measured-execution/inventory
-> (token-resolution | dex-api-token-pricing) -> dex-liquidity/types. The odd
edge was the core types module reaching into the feature module for the
UniV3ExecutionCandidate shape. That type now lives in a leaf module
(measured-execution/candidate-types.ts, zero imports); inventory re-exports it
for existing consumers, types.ts points its inline type import at the leaf, and
inventory aliases canonicalExitRouteAssetKey directly instead of importing
token-resolution's one-line wrapper. No behavior change; cycle check clean;
process-pools suites 20/20.
